do it fast

+++ bin/cleancluster.sh
@@ -1,6 +1,11 @@
 source nesox.sh
 
+if [ "$1" = "" ]
+then
 command="rm -rf $logs/*.log"
+else
+command="rm -rf $logs/nesox-reader-*.log"
+fi
 
 for node in $nodes
 do
